+++ src/tests/sqlsmith/src/runner.rs
@@ -14,10 +14,11 @@
 
 use itertools::Itertools;
 use rand::{Rng, SeedableRng};
-use risingwave_expr::error::ExprError;
 use tokio_postgres::error::Error as PgError;
 
-use crate::{create_table_statement_to_table, mview_sql_gen, parse_sql, sql_gen, Table};
+use crate::{
+    create_table_statement_to_table, is_permissible_error, mview_sql_gen, parse_sql, sql_gen, Table,
+};
 
 /// e2e test runner for sqlsmith
 pub async fn run(client: &tokio_postgres::Client, testdata: &str, count: usize) {
@@ -169,28 +170,6 @@ async fn drop_tables(mviews: &[Table], testdata: &str, client: &tokio_postgres::
     for stmt in sql.lines() {
         client.execute(stmt, &[]).await.unwrap();
     }
-}
-
-fn is_division_by_zero_err(db_error: &str) -> bool {
-    db_error.contains(&ExprError::DivisionByZero.to_string())
-}
-
-fn is_numeric_out_of_range_err(db_error: &str) -> bool {
-    db_error.contains(&ExprError::NumericOutOfRange.to_string())
-}
-
-/// Skip queries with unimplemented features
-fn is_unimplemented_error(db_error: &str) -> bool {
-    db_error.contains("Feature is not yet implemented")
-}
-
-/// Certain errors are permitted to occur. This is because:
-/// 1. It is more complex to generate queries without these errors.
-/// 2. These errors seldom occur, skipping them won't affect overall effectiveness of sqlsmith.
-fn is_permissible_error(db_error: &str) -> bool {
-    is_numeric_out_of_range_err(db_error)
-        || is_division_by_zero_err(db_error)
-        || is_unimplemented_error(db_error)
 }
 
 /// Validate client responses
